Refactorizando front, hook personalizado, división de código en componentes
