+++ frontend/components/Calculator.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import Button from "./Button";
-import Input from "./Input";
+import Button from "./ui/Button";
+import Input from "./ui/Input";
 
 interface CalculatorProps {
   devolverResultado: (estimacion: number) => void;
