+++ frontend/components/CalculatorForm.tsx
@@ -1,5 +1,5 @@
-import Button from "./Button";
-import Input from "./Input";
+import Button from "./ui/Button";
+import Input from "./ui/Input";
 
 
 export default function CalculatorForm() {
