+++ frontend/components/__tests__/Button.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, fireEvent } from "@testing-library/react";
-import Button from "../Button";
+import Button from "../ui/Button";
 
 describe("Button", () => {
   test("renderiza el título recibido por props", () => {
