+++ frontend/components/tareas/Card.tsx
