+++ frontend/components/ui/Error.tsx
