+++ frontend/pages/admin.tsx
@@ -1,106 +1,30 @@
-import CardProyecto from "@/components/CardProyecto";
-import MensajeError from "@/components/Error";
-import { ApiGetProyecto, ApiCrearProyecto, ApiCrearTareasBulk, ApiGetTareasProyecto } from "@/utils/api";
-import { parsearArchivoTareas } from "@/utils/parsearArchivo";
+import CardProyecto from "@/components/admin/CardProyecto";
+import MensajeError from "@/components/ui/Error";
+import ProyectoForm from "@/components/admin/ProyectoForm";
+import Success from "@/components/ui/Success";
+import { useProyectos } from "@/hooks/useProyectos";
 import { useUserContext } from "@/utils/userContext";
-import { useEffect, useState } from "react";
+import {  useState } from "react";
 
 type ErrorFila = { fila: number; mensajes: string[] };
 
-type Usuario = {
-  id: number
-  nombre: string
-  apellido: string
-  email: string
-  rol_admin: boolean
-}
-
-type Proyecto = {
-  id: number
-  titulo: string
-  descripcion: string
-  fechaCreacion: string
-  lider: Usuario
-}
-
 export default function Admin() {
-  const [proyectos, setProyectos] = useState<Proyecto[]>([]);
   const [show, setShow] = useState(false);
   const { user, token } = useUserContext();
   const [error, setError] = useState("");
-  const [titulo, setTitulo] = useState("");
-  const [descripcion, setDescripcion] = useState("");
-  const [file, setFile] = useState<File | null>(null);
+  const [success, setSuccess] = useState("");
   const [erroresFilas, setErroresFilas] = useState<ErrorFila[]>([]);
 
-  useEffect(() => {
-    if (!user?.id) return;
-    async function obtenerProyecto() {
-      try {
-        if (!user) return;
-        const response = await ApiGetProyecto(user.id, token);
-        if (!response.ok) {
-          setError("No se pudieron obtener los proyectos.");
-          return;
-        }
-        const data = await response.json();
-        setProyectos(data);
-      } catch (err) {
-        console.error(err);
-      }
-    }
-    obtenerProyecto();
-  }, [user, token]);
+  const { proyectos, setProyectos } = useProyectos(user?.id, token);
 
   function handleCreate() {
     setShow(true);
   }
 
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setError("");
-    setErroresFilas([]);
-
-    try {
-      if (!user) return;
-      const resProyecto = await ApiCrearProyecto(
-        { titulo, descripcion, idLider: Number(user.id) },
-        token,
-      );
-      if (!resProyecto.ok) {
-        const err = await resProyecto.json().catch(() => null);
-        const msg = err?.message;
-        setError(Array.isArray(msg) ? msg.join(" — ") : msg ?? "No se pudo crear el proyecto.");
-        return;
-      }
-      const proyecto = await resProyecto.json();
-
-      if (file) {
-        const tareas = await parsearArchivoTareas(file);
-        const resBulk = await ApiCrearTareasBulk({ idProyecto: proyecto.id, tareas }, token);
-
-        if (!resBulk.ok) {
-          const detalle = await resBulk.json();
-          setError(detalle.mensaje ?? "Hay filas con errores en el archivo.");
-          setErroresFilas(detalle.errores ?? []);
-          return;
-        }
-      }
-
-      setProyectos((prev) => [...prev, proyecto]);
-      setShow(false);
-      setTitulo("");
-      setDescripcion("");
-      setFile(null);
-    } catch (err) {
-      console.error(err);
-      setError("Ocurrió un error procesando el archivo. Volvé a seleccionarlo y reintentá.");
-    }
-  }
-
   return (
     <>
       {error && <MensajeError text={error} />}
+      {success && <Success text={success} />}
 
       {erroresFilas.length > 0 && (
         <div className="mx-20 mt-6 rounded-2xl border border-red-300 bg-red-50 p-4">
@@ -126,39 +50,19 @@ export default function Admin() {
           </button>
         </div>
         {show ? (
-          <div className="flex border-1 border-gray-800 rounded-2xl p-6 shadow-xl">
-            <form className="flex flex-col w-full space-y-3" onSubmit={handleSubmit}>
-              <label>Título:</label>
-              <input
-                className="bg-white rounded-md p-2 border-1 border-secondary shadow-xl"
-                type="text" placeholder="Título"
-                value={titulo} onChange={(e) => setTitulo(e.target.value)}
-              />
-              <label>Descripción:</label>
-              <textarea
-                className="bg-white rounded-md p-2 border-1 border-secondary shadow-xl"
-                placeholder="Descripción"
-                value={descripcion} onChange={(e) => setDescripcion(e.target.value)}
-              />
-              <label>Subir archivo (.xls / .csv)</label>
-              <input
-                className="bg-white rounded-md p-2 border-1 border-secondary shadow-xl"
-                type="file" accept=".xls,.xlsx,.csv"
-                onChange={(e) => setFile(e.target.files?.[0] ?? null)}
-              />
-              <button
-                type="submit"
-                className="flex justify-center mt-4 w-[200px] rounded-full bg-secondary px-3 py-1.5 font-semibold text-white cursor-pointer hover:bg-blue-400"
-              >
-                Guardar
-              </button>
-            </form>
-          </div>
+          <ProyectoForm
+            onProyectoCreado={(nuevoProyecto) =>
+              setProyectos((prev) => [...prev, nuevoProyecto])
+            }
+            onClose={() => setShow(false)}
+            setError={setError}
+            setErroresFilas={setErroresFilas}
+          />
         ) : (
           <></>
         )}
         {proyectos.map((proyecto) => (
-          <CardProyecto proyecto={proyecto} key={proyecto.id} />
+          <CardProyecto proyecto={proyecto} setError={setError} key={proyecto.id} setSuccess={setSuccess} />
         ))}
       </div>
     </>
